backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
